# optimize Vector{T} constuctor by avoiding elmlist variable