The `elmlist` variable was storing either the `ElmList` or the `Globals.ELM_LIST`
object, and was called later as `elmlist(obj, i)`. But unfortunately,
this is not (yet) a pattern that the Julia compiler optimizes well.
So instead, just write explicitly the calls to `ElmList` and
`ELM_LIST`, with a branch.

The timings for
```julia
lj = rand(1:99, 999); l = GAP.julia_to_gap(lj); @btime Vector{Int}($l);
```
look like
```
  337.507 μs (1982 allocations: 58.66 KiB) # master
  102.435 μs (1494 allocations: 35.42 KiB) # PR
  46.805 μs (980 allocations: 23.56 KiB) # master + PR 539
  31.162 μs (1491 allocations: 31.55 KiB)  # PR + PR 539
```
Together with PR 539, this is still an improvement, but involves more allocations, I'm not sure why. By putting the `if islist` test out of the loop (meaning duplicating the loop), the extra allocations disappear, and we get 
```
26.122 μs (492 allocations: 15.94 KiB)
```
But it's annoying to duplicate code like that, so I'd rather not do so too hastily.