@ThomasBreuer 2020-10-20
This is interesting.
The point seems to be that `GAP.ElmList` is a Julia function whereas `GAP.Globals.ELM_LIST` is a `GAP.ForeignGAP.MPtr` object.
(I have looked at related places in the code, but I did not find other instances of this situation.)