@rfourquet 2020-10-20
> The point seems to be that GAP.ElmList is a Julia function whereas GAP.Globals.ELM_LIST is a GAP.ForeignGAP.MPtr object.

Yes; if both were functions, the situation would be the same because each function has its own type. But I agree it's tempting to factor out the code by putting the function in a variable; I'm not clear on whether this is an optimization that we can hope the compiler to implement any time soon...